tools/global/get_data_registry_dataset/custom_model/requirements.txt
opentelemetry-instrumentation-requests==0.54b1
pyarrow==20.0.0
